==> Preprocess/step0_preprocess.ipynb <==
# change the file path here

RAMDOMSEED = 11

# original data download from BindingDB
datafile = "/home/wbm001/deeplpi/DeepLPI/data/BindingDB_All.tsv"
datapath = "/home/wbm001/deeplpi/DeepLPI/data/"

import pandas as pd

data = pd.read_csv(datafile,'\t',error_bad_lines=False)

oridata = data.copy()

data

print("mollen: " + str(len(set(data["Ligand SMILES"]))))
print("seqlen: " + str(len(set(data["BindingDB Target Chain  Sequence"]))))

# data = oridata.copy()

import numpy as np

data = data.set_index("BindingDB Reactant_set_id")
data = data.drop(data["BindingDB Target Chain  Sequence.1"].dropna().index) # remove multichain sequence
print("after remove multichain sequence: " + str(len(data)))
print("mollen: " + str(len(set(data["Ligand SMILES"]))))
print("seqlen: " + str(len(set(data["BindingDB Target Chain  Sequence"]))))
data = data[["Ligand SMILES","Ki (nM)","Kd (nM)","BindingDB Target Chain  Sequence"]]
dataki = data[["Ligand SMILES","Ki (nM)","BindingDB Target Chain  Sequence"]].dropna() # remove data without ki and kd value
datakd = data[["Ligand SMILES","Kd (nM)","BindingDB Target Chain  Sequence"]].dropna()
print("data have ki value: " + str(len(dataki)))
print("data have kd value: " + str(len(datakd)))
datakd

print("mollen: " + str(len(set(datakd["Ligand SMILES"]))))
print("seqlen: " + str(len(set(datakd["BindingDB Target Chain  Sequence"]))))

from math import log10


dataki["Ki (nM)"] = dataki["Ki (nM)"].map(lambda x : float(x) if str(x)[0] not in [">","<"] else float(str(x)[1:]))
datakd["Kd (nM)"] = datakd["Kd (nM)"].map(lambda x : float(x) if str(x)[0] not in [">","<"] else float(str(x)[1:]))

dataki["pki"] = dataki["Ki (nM)"].map(lambda x : -log10(x) if x != 0 else None)
datakd["pkd"] = datakd["Kd (nM)"].map(lambda x : -log10(x) if x != 0 else None)

# remove unproparate ki | kd

dataki = dataki.dropna()
datakd = datakd.dropna()

print("data ki remain: " + str(len(dataki)))
print("data kd remain: " + str(len(datakd)))

print("mollen: " + str(len(set(datakd["Ligand SMILES"]))))
print("seqlen: " + str(len(set(datakd["BindingDB Target Chain  Sequence"]))))

# select out the test mol and seq

from sklearn.model_selection import train_test_split

_, ki_mol_test = train_test_split(list(dataki["Ligand SMILES"]), test_size=700, random_state=RAMDOMSEED)
_, ki_seq_test = train_test_split(list(dataki["BindingDB Target Chain  Sequence"]), test_size=6, random_state=RAMDOMSEED)
_, kd_mol_test = train_test_split(list(datakd["Ligand SMILES"]), test_size=50, random_state=RAMDOMSEED)
_, kd_seq_test = train_test_split(list(datakd["BindingDB Target Chain  Sequence"]), test_size=15, random_state=RAMDOMSEED)

molset = set(list(dataki["Ligand SMILES"]) + list(datakd["Ligand SMILES"]))
moldict = dict(enumerate(molset))
molrev = { y:x for x,y in enumerate(molset) }

seqset = set(list(dataki["BindingDB Target Chain  Sequence"]) + list(datakd["BindingDB Target Chain  Sequence"]))
seqdict = dict(enumerate(seqset))
seqrev = { y:x for x,y in enumerate(seqset) }

moldict = pd.DataFrame.from_dict(moldict,orient='index')
seqdict = pd.DataFrame.from_dict(seqdict,orient='index')

moldict["ki_test"] = moldict[0].map(lambda x : True if x in ki_mol_test else False)
seqdict["ki_test"] = seqdict[0].map(lambda x : True if x in ki_seq_test else False)
moldict["kd_test"] = moldict[0].map(lambda x : True if x in kd_mol_test else False)
seqdict["kd_test"] = seqdict[0].map(lambda x : True if x in kd_seq_test else False)

dataki["mol"] = dataki["Ligand SMILES"].map(lambda x : molrev[x])
datakd["mol"] = datakd["Ligand SMILES"].map(lambda x : molrev[x])

dataki["seq"] = dataki["BindingDB Target Chain  Sequence"].map(lambda x : seqrev[x])
datakd["seq"] = datakd["BindingDB Target Chain  Sequence"].map(lambda x : seqrev[x])

# drop duplicates
dataki = dataki.drop_duplicates(subset=["mol","seq"],keep='first')
datakd = datakd.drop_duplicates(subset=["mol","seq"],keep='first')
print("data have ki remain: " + str(len(dataki)))
print("data have kd remain: " + str(len(datakd)))
print("mollen: " + str(len(set(datakd["Ligand SMILES"]))))
print("seqlen: " + str(len(set(datakd["BindingDB Target Chain  Sequence"]))))

dataki = dataki[["mol","seq","pki"]]
dataki["state"] = 0
datakd = datakd[["mol","seq","pkd"]]
datakd["state"] = 0

dataki["state"] += dataki["mol"].map(lambda x : 1 if moldict.loc[x]["ki_test"] else 0)
dataki["state"] += dataki["seq"].map(lambda x : 2 if seqdict.loc[x]["ki_test"] else 0)
datakd["state"] += datakd["mol"].map(lambda x : 1 if moldict.loc[x]["kd_test"] else 0)
datakd["state"] += datakd["seq"].map(lambda x : 2 if seqdict.loc[x]["kd_test"] else 0)

dataki["dup"] = dataki.duplicated(subset=["mol"],keep='first') & dataki.duplicated(subset=["seq"],keep='first')
_,ki_both_test = train_test_split(np.array(dataki.loc[dataki["dup"]==True].loc[dataki["state"]==0].index), test_size=1200, random_state=RAMDOMSEED)

datakd["dup"] = datakd.duplicated(subset=["mol"],keep='first') & datakd.duplicated(subset=["seq"],keep='first')
_,kd_both_test = train_test_split(np.array(datakd.loc[datakd["dup"]==True].loc[datakd["state"]==0].index), test_size=2500, random_state=RAMDOMSEED)

dataki["state"] += dataki.index.map(lambda x : 10 if x in ki_both_test else 0)
datakd["state"] += datakd.index.map(lambda x : 10 if x in kd_both_test else 0)

ki_train = dataki.loc[dataki["state"]==0]
ki_test = dataki.loc[dataki["state"]!=0]
kd_train = datakd.loc[datakd["state"]==0]
kd_test = datakd.loc[datakd["state"]!=0]

ki_test["unseen"] = ki_test["state"].map(lambda x : "mol" if x == 1 else "seq" if x == 2 else "none" if x == 10 else "both")
kd_test["unseen"] = kd_test["state"].map(lambda x : "mol" if x == 1 else "seq" if x == 2 else "none" if x == 10 else "both")

ki_train[["mol","seq","pki"]].to_csv(datapath + "ki_train.csv")
kd_train[["mol","seq","pkd"]].to_csv(datapath + "kd_train.csv")
ki_test[["mol","seq","pki","unseen"]].to_csv(datapath + "ki_test.csv")
kd_test[["mol","seq","pkd","unseen"]].to_csv(datapath + "kd_test.csv")

moldict.to_csv(datapath + "mol_dict.csv")
seqdict.to_csv(datapath + "seq_dict.csv")
