==> Preprocess/step1_molembedding.ipynb <==
datapath = "/home/wbm001/deeplpi/DeepLPI/data/"
datafile = '/home/wbm001/deeplpi/DeepLPI/data/mol_dict.csv'
modelpath = "/home/wbm001/deeplpi/DeepLPI/preprocess/mol2vec/model_300dim.pkl"

from rdkit.Chem import SDMolSupplier
from rdkit.Chem.Draw import IPythonConsole

import pandas as pd

data = pd.read_csv(datafile)

data = data[["0"]]
data

from mol2vec.features import mol2alt_sentence, MolSentence, DfVec, sentences2vec
from gensim.models import Word2Vec
from rdkit.Chem import MolFromSmiles as sm
import numpy as np

data["sent"] = data["0"].map(lambda x : sm(x))

def mol2alt(x):
    try:
        re = mol2alt_sentence(x,1)
    except:
        re = None
    return re

data

data_sentence = data["sent"].map(lambda x : mol2alt(x))

data["alt_sent"] = data_sentence
data

from gensim.models import word2vec

model = word2vec.Word2Vec.load(modelpath)

data = data.dropna()
data

molvec = sentences2vec(data["alt_sent"],model,unseen='UNK')
molvec

molembed = np.hstack((np.array(data.index).reshape(-1,1) , molvec))

molembeddf = pd.DataFrame(molembed)
molembeddf[0] = molembeddf[0].astype(int)
molembeddf = molembeddf.set_index(0)
molembeddf

molembeddf.to_csv(datapath+"mol_embed.csv")
